src/python/lib/lib_a/pyproject.toml
[tool.poetry]
name = "lib_a"
version = "0.1.0"
description = ""
authors = ["peachanG <kenkman0427@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.20.1"
absl-py = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
